refactor: pull search bar into own component
