+++ src/routes/sidebar.svelte
@@ -1,0 +1,8 @@
+
+<script lang="ts">
+
+</script>
+
+<div>
+  
+</div>
